withdraw WETH test

+++ contracts/bridge/interfaces/IWETH9.sol
@@ -9,25 +9,25 @@ interface IWETH9 {
 
     function decimals() external view returns (uint8);
 
-    function balanceOf(address) external view returns (uint256);
+    function balanceOf(address) external view returns (uint);
 
-    function allowance(address, address) external view returns (uint256);
+    function allowance(address, address) external view returns (uint);
 
     receive() external payable;
 
     function deposit() external payable;
 
-    function withdraw(uint256 wad) external;
+    function withdraw(uint wad) external;
 
     function totalSupply() external view returns (uint256);
 
-    function approve(address guy, uint256 wad) external returns (bool);
+    function approve(address guy, uint wad) external returns (bool);
 
-    function transfer(address dst, uint256 wad) external returns (bool);
+    function transfer(address dst, uint wad) external returns (bool);
 
     function transferFrom(
         address src,
         address dst,
-        uint256 wad
+        uint wad
     ) external returns (bool);
 }
